+++ tests/test_confirmable.py
@@ -139,6 +139,9 @@ def test_confirmable_flag(app, clients, get_message):
 
 
 @pytest.mark.registerable()
+@pytest.mark.settings(
+    email_template_confirm="confirmation_instructions_test", email_html=False
+)
 def test_confirmation_template(app, client, get_message, outbox):
     # Check contents of email template - this uses a test template
     # in order to check all context vars since the default template
@@ -159,10 +162,7 @@ def test_confirmation_template(app, client, get_message, outbox):
         )
         # Register - this will use the welcome template
         client.post("/register", data=data, follow_redirects=True)
-        # Explicitly ask for confirmation -
-        # this will use the confirmation_instructions template
-        client.post("/confirm", data=dict(email="mary@lp.com"))
-        assert len(outbox) == 2
+        assert len(outbox) == 1
         # check registration email
         matcher = re.findall(r"\w+:.*", outbox[0].body, re.IGNORECASE)
         # should be 4 - link, email, token, config item
@@ -171,19 +171,26 @@ def test_confirmation_template(app, client, get_message, outbox):
         assert matcher[2].split(":")[1] == registrations[0]["confirmation_token"]
         assert matcher[3].split(":")[1] == "True"  # register_blueprint
 
+        # Explicitly ask for confirmation -
+        # this will use the confirmation_instructions template
+        client.post("/confirm", data=dict(email="mary@lp.com"))
         # check confirmation email
         matcher = re.findall(r"\w+:.*", outbox[1].body, re.IGNORECASE)
-        # should be 4 - link, email, token, config item
+        # should be 5 - link, email, token, config item, within
         assert matcher[1].split(":")[1] == "mary@lp.com"
         assert matcher[2].split(":")[1] == recorded_tokens_sent[0]
         token = matcher[2].split(":")[1]
         assert token == recorded_tokens_sent[0]
         assert matcher[3].split(":")[1] == "True"  # register_blueprint
+        assert matcher[4].split(":")[1] == "2 days"  # within
 
         # check link
         _, link = matcher[0].split(":", 1)
         response = client.get(link, follow_redirects=True)
         assert get_message("EMAIL_CONFIRMED") in response.data
+
+        # regular text template is included in the test template
+        assert "This link will expire in 2 days." in outbox[1].body
 
 
 @pytest.mark.registerable()